feat: show actual error message in tenant middleware
Still getting 500 errors but can't see what's actually failing.

Added errorMessage to the response so we can see the actual error.
This will help us understand why tenant resolution is failing.

The error is being caught but we don't know if it's:
- Database connection issue
- Prisma query failing
- Tenant not found
- Something else

After deployment, the error response will include the actual error message.